watchdog: imx2_wdt: shorten watchdog timer at shutdown
Some driver on the RC board is intermittently preventing
a clean shutdown and triggering the watchdog at its'
maximum duration.

Reduce the duration to 5 seconds.

This is safe because the persistent storage on the device
is SD card and the SD card interface is shut down before
we get to the watchdog driver.

Signed-off-by: Eric Nelson <eric@nelint.com>